Added Day 624 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
In 2000, the digital world encountered a new threat: the Distributed Denial of Service (DDoS) attack. Internet giants like Yahoo, eBay, and Amazon were crippled by a flood of traffic from numerous sources, orchestrated by a 15-year-old hacker known as 'Mafiaboy', real-life Michael Calce from Canada. This revolutionary type of attack managed to halt the services of the strongest online platforms of the time.
Today, Michael Calce is not only a respected white-hat hacker but also a legend of cyberspace.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
У 2000 році цифровий світ зіткнувся з новим видом загрози: атакою типу "розподілена відмова в обслуговуванні" (DDoS). Гіганти інтернету як Yahoo, eBay та Amazon були паралізовані потоком інтернет-трафіку з багатьох джерел, організованим 15-річним хакером на прізвисько 'Mafiaboy', в житті Майкл Калс з Канади. Цим революційно новим типом атаки вдалося зупинити сервіси найміцніших онлайн-платформ того часу. 
Сьогодні Майкл Калс це не тільки поважний білий хакер, але і легенда кіберпростору.