Added Day 627 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
In case you missed it, here's an inspiring story about a white-hat hacker "blazezaria", who earned nearly $1 million through the bug bounty platform HackenProof for discovering vulnerabilities and donated 10% to the "Return Alive" charity fund.
We're helping not only in cyberspace! Continue to support the Armed Forces of Ukraine through verified funds like UNITED24 and Back Alive.
Got your own story or achievement? Share it with us by sending it to armyuait@gmail.com, and we might feature it on our channel!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Якщо ви пропустили, ось мотивуюча історія про білого хакера "blazezaria", який через багбаунті-платформу HackenProof заробив майже 1 мільйон доларів за виявлені вразливості та пожертвував 10% в благодійний фонд "Повернись живим".
Ми допомагаємо не лише у кіберпросторі! Продовжуйте підтримувати ЗСУ через перевірені фонди UNITED24 та Повернись Живим.
Маєте свою історію чи досягнення? Поділіться з нами, надіславши їх на електронну адресу armyuait@gmail.com, і ми опублікуємо їх на нашому каналі!