Beta ➜ Main | Added Day 628 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
What if our world is more ordered than it appears? Melvin Vopson has introduced the second law of infodynamics, suggesting that both our genetics and the digital world may naturally incline towards order rather than chaos, as previously thought. It's as if your room cleans itself up. And if that's the case, could we be living in some sort of grand computer simulation, like "The Matrix"? It's quite unconventional, but Vopson gives us scientific reasons to ponder. It sounds incredible, but it's real science!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
А що якщо наш світ більш упорядкований, ніж здається? Мелвін Вопсон запропонував другий закон інфодинаміки, який показує, що і наша генетика, і цифровий світ можуть тяжіти до порядку, а не хаосу, як вважалося раніше. Це наче ваша кімната сама себе прибирає. І якщо це так, то чи живемо ми в якомусь великому комп'ютерному симуляторі, типу "Матриці"? Дуже незвично, але Вопсон дає нам наукові підстави для роздумів. Це звучить неймовірно, але це реальна наука!