Added Day 635 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
"NCSC-WA-002-85," also known as "Personal Computer Security Considerations," originated from the National Computer Security Center, a part of the United States Department of Defense. Published on December 1, 1985, this document was an early recognition by government entities of the increasing significance of personal computers in both business and personal sectors, and the potential security threats that could arise with their widespread adoption. They were the oracles of their time—no more, no less.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
"NCSC-WA-002-85", також відомий як "Розгляд питань безпеки персональних комп'ютерів", походить з Національного центру комп'ютерної безпеки, що є частиною Міністерства оборони Сполучених Штатів. Опублікований 1 грудня 1985 року, цей документ був раннім офіційним визнанням зростального значення персональних комп'ютерів у бізнесі та особистому секторі, а також потенційних загроз кібербезпеки, які могли виникнути у зв'язку з їх поширенням. Пророцтво - не більше, не менше.