Added Day 636 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
This time, we successfully disconnected the major internet provider Trinity in Donetsk, which is also used by the enemy for military purposes. This is the result of our consistent strategy aimed at complicating the enemy's use of any infrastructure.
For those who still doubt how significant a contribution the IT ARMY is making to the enemy's cyberspace - read the Kommersant article, where telecom providers have become the most popular targets in the Runet in the third quarter. A coincidence? Definitely not.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Цього разу, нам вдалося відключити великого інтернет-провайдера Trinity у Донецьку, який також використовується ворогом у військових цілях. Це результат нашої послідовної стратегії спрямованої на ускладнення використання ворогом будь-якої інфраструктури. 
Для тих хто досі сумнівається наскільки великий внесок робить IT ARMY у ворожий кібернетичний простір - читайте статтю Комерсант, де телеком провайдери стали найпопулярнішими цілями в рунеті в третьому кварталі. Збіг? Ось точно ніт.