Added Day 637 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
We can now talk about our operation against another major provider, Donapex in Donetsk, which started a week ago. It's regrettable that our citizens remaining in the occupied territories also suffer from this, but unfortunately, in the modern world, almost all infrastructure is used by both civilians and the military.
The struggle continues and we all must only intensify our efforts! Join our operations; it is possible with any device.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Тепер вже можна розповісти й про нашу операцію проти іншого великого провайдера Донапекс в Донецьку, яка почалась ще тиждень тому. Шкода що від цього також страждають наші громадяни, які залишились на окупованих територіях, але в сучасному світі, на жаль, майже вся інфраструктура використовується як цивільними, так і військовими.
Боротьба триває і ми всі маємо тільки нарощувати наші зусилля! Приєднуйтесь до наших операцій, це можливо з будь-якими девайсами.