Added Day 640 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Our work becomes an art! Just look at the beauty our volunteers are assembling for optimizing the cost of hardware and power for attacks. A little more, and creating a wall for DDOS at home and installing our software will be no more complicated than setting up a robot vacuum cleaner.
Want such hardware? - ask the volunteers in our chat. Meanwhile, there are just a few days left until the big release of the new cross-platform application from the IT ARMY, which will be incredibly easy to install on all non-Linux platforms.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Робота IT ARMY стає мистецтвом! Тільки подивіться яку красу збирають наші волонтери для оптимізації вартості заліза і потужностей для атак. Ще трошки та створення стінки для DDOS вдома і встановлення нашого ПО стане не складніше ніж налаштування робота-пилососа.
Хочете таке залізо? - запитайте волонтерів в нашому чаті. А тим часом залишилось декілька днів до великого релізу нового кросплатформного застосунку від IT ARMY який буде неймовірно легко встановити на всі не-Лінукс платформи.