Added Day 646 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Another provider, calling itself Obltelecom, in the enemy's capital, has fallen under our pressure. For the second month now, we're focusing on communications and internet infrastructure, and although our real targets will never write about it online - we know they're suffering.
Maybe we're less diverse now, but much more systematic and consistent. Do you support such a tactic of the IT ARMY? Write in the chat.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Черговий провайдер, що кличе себе Облтелеком, у столиці ворога впав під нашим натиском. Вже другий місяць ми фокусуємось саме на зв'язку та інтернет інфраструктурі, і хоча наші справжні цілі ніколи не напишуть про це в інтернеті - ми знаємо що вони страждають. 
Можливо ми зараз менш різноманітні, але набагато більш системні й послідовні. Чи підтримуєте ви таку тактику IT ARMY? Напишіть у чаті.