Added Day 651 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Here is Mariino, not like in Donetsk region, but another internet provider in Moscow that we disconnected as part of the enemy de-internetization program. We didn't find any complaints this time, but believe me, the provider is down 'as it should be'.
We remind you that interruptions in the internet significantly affect the efficiency not only of government bodies, but also of the military and special services. Every day, every hour they could not work, goes into our victory fund. So, we grit our teeth and keep working. Against all odds!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Ось і Мар'їно, не те що на Донеччині, а черговий інтернет-провайдер в Москві, який ми відключили в межах програми деінтернетізації ворога. Скарг цього разу не знайшли, але повірте, провайдер лежить "як треба".
Ще раз нагадуємо що перебої з інтернетом суттєво впливають на працездатність не тільки органів влади, але і військових та спецслужб. Кожен день, кожна година яку вони не мали відпрацювати, йде до нашої скрині до перемоги. Тому стискаємо зуби й працюємо. Попре все!