Added Day 659 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
📢 Hot Fix for IT ARMY Kit! 🚀‍
‍- MHDDoS temporarily removed from macOS due to compatibility issues
- Db1000n now excluded from random selection, can only be enabled manually
- Fixed the issue of MHDDoS process duplication
- Standardized MHDDoS parameters for more efficient operation
- Changing modules automatically downloads and runs the latest version
- All logs now consolidated in one window with auto-scroll‍
Get ready for a new, more powerful, and user-friendly IT ARMY Kit! 🛠️💪💻
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
📢  Швидкий патч IT ARMY Kit! 🚀
- MHDDoS поки прибрано з macOS через несумісність
- Db1000n тепер прибрано з рандомного вибору і може бути включений тільки вручну
- Усунено проблему з подвоєнням процесів MHDDoS
- Стандартизовано параметри MHDDoS для більш ефективної роботи
- Зміна модуля автоматично завантажує та запускає останню версію
- Всі журнали тепер зібрані в одному вікні з автопрокруткою
Готуйтеся до нового, потужнішого та зручнішого IT ARMY Kit! 🛠️💪💻