Added Day 667 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Yesterday, Bitrix failed to fend off our attack, and it continues today. Two days are good, but now a third day has begun... You guys are doing great, every little device creates the effect of a mosquito fleet, which is so important for the effectiveness of the attack.‍
By the way, Bitrix is 'heroically' talking about temporary technical issues. They might as well say that's how it's supposed to work 😂😂😂‍
‍Shall we hold them off for another day, cyber resistance?
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
За вчора Бітріксу не вдалося відбитися від нашої атаки й вона продовжується сьогодні. Дві доби добре, але ось і третя пішла... Ви молодці, кожен навіть маленький девайс робить ефект москітного флоту тому це так важливо для ефективності атак.
Бітрікс, до речі, "героїчно" розповідає про тимчасові технічні неполадки. Вже сказали б що воно так і має працювати 😂😂😂 
Втримаємо них ще день, кіберспротив? 