Added Day 775 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
The federal internet provider JustLan might as well be renamed to JustDead. It would sound cooler and match the current state.
This is significant. Taking down such providers is not only crucial for crashing a huge number of organizations that depend on them but also clearly demonstrates our growing capabilities.
Friends, we constantly note the increasing number of resources you allocate for the needs of the IT ARMY. For D-Day, we will need everything. Don't stop!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Федерального інтернет-провайдера JustLan можна перейменовувати на JustDead. Буде і звучати крутіше і відповідати поточному стану. 
Це вже масштаб. Вибивання таких провайдерів не тільки важливо для обвалення величезної кількості організацій, які на них зав'язані, але і чітко показують наші зростаючі можливості. 
Друзі, ми відмічаємо постійно зростаючу кількість ресурсів, яку ви залучаєте під потреби IT ARMY. Для дня Ч нам знадобиться все. Не зупинятесь!