Added Day 778 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Our enemy has long been without a system of justice; it's unnecessary in authoritarian states where the law is always used for political purposes. The enemy country is no exception. Mass repressions against anyone who could remotely be called opposition prove this.
Therefore, courts become solely a tool for resolving economic disputes, and thus a legitimate target for the IT ARMY.
This means we're disabling the case management system for the majority of Russian courts sudrf.ru and Mosgorsud.
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
В нашого ворога вже давно нема системи правосуддя, вона не потрібна в авторитарних державах, де закон завжди використовується у політичних цілях. Країна-ворог не виняток. Масові репресії проти всіх тих кого хоч якось можна було називати опозицією є тому підтвердженням. 
Тому суди стають виключно інструментом розв'язання економічних суперечок, а відповідно законною ціллю для IT ARMY.
Значить відключаємо систему управління справами для більшості російських судів sudrf.ru та Мосгорсуд.