Added Day 783 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
The Iranian hacker group Handala has claimed that they successfully attacked the Israeli Iron Dome missile defense system, according to the cyber hackers' Telegram channel," was widely reported across the Russian internet. Currently, we know that the "disabled" Israeli air defense system shot down almost everything that was flying.
This is a classic example of psychological operations in modern wars and another reminder that not only anonymous Telegram channels but also low-quality media are not reliable sources of information. Stay vigilant!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Иранская хакерская группировка Handala заявила, что провела успешную атаку на израильскую систему противоракетной обороны (ПРО) «Железный купол», сообщается в Telegram-канале кибервзломщиков." було розтиражовано по всьому рунету. На зараз ми знаємо що "відключене" ізраїльське ППО збило майже все що летіло.
Це класичний випадок ІПСО сучасних війн і чергове нагадування, що не тільки анонімні Телеграм канали, а низькосортне медіа не є джерелом надійної інформації. Будьте пильні!