Added Day 784 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Mosnet provider is our cyber-Chornobaivka, friends. Well, you get it...
But it's rare that us addressed in official press releases, which can be briefly summarized as: 'We are not into politics. What for?'. Our response: For not learning to control your own power, for your power starting a war and only a small part of the evil you brought to our land has reached you so far. Eat, don't make a mess.
A few more satellite internet providers have also gone offline. What a day!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про цілі:
Провайдер Моснет це наша кібер-Чорнобаївка, друзі. Ну ви зрозуміли... 
Але нечасто до нас звертаються в офіційних прес-релізах наші цілі, який можна коротко підсумувати як: "Ми внє палітікі. За що?". Наша відповідь: За те що не навчились контролювати власну владу, за те що ваша влада розпочала війну і до вас поки дійшла тільки маленька частка того лиха яке ви принесли на нашу землю. Їжте, не обляпайтесь.
Ще декілька провайдерів супутникового інтернету також пішли в офлайн. Оце так день!