Added Day 787 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
While the enemy spends billions of dollars building their own satellite internet network, we have our response ready—two of the largest providers, Astra and Allegrosky, have been down for a few days in row.
The article also complains that, like StarLink, they somehow can't make it cheap. It's strange, why is that? Maybe because instead of focusing on reducing the cost of space travel, they are busy restoring T-55 tanks?
Launch the satellites, we are waiting!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Щоб не відставати від цілей, ми збираємося використовувати цілі db1000n безпосередньо з їхнього сховища GitHub.
Ці оновлення відбуватимуться щодня, тож якщо db1000n зміниться, ймовірно, знадобиться кілька годин більше, перш ніж зміни з’являться тут.
Інформація про ціл